C++: compiler detection and build command dependency.
The C++ compiler in use is now detected; vendor-specific compiler flags
are now supported for C++ files.

C++ file's output now also depends on the compiler (so if you update the
compiler, everything will be rebuilt) and on the exact command line used
to build it; this, together with the fact that $CXXFLAGS is used in the
command line, means that changing that environment variable prompts a
rebuild.

At some point in the future I'll probably introduce a mode that fixes
the contents of CXXFLAGS from the first build that happens for an output
directory, plus some additional caching (like not having to respecify
the compiler via CXX), but it's not a priority right now.